Formatted Code Structure and Changed Minimap position
