+++ frontend/src/css/card.module.css
@@ -11,6 +11,7 @@
   text-align: center;
   padding: 0.5rem;
   border-bottom: 0.1rem #000000 solid;
+  font-size: medium;
 }
 
 .cardBody {
@@ -28,6 +29,8 @@
   flex-wrap: wrap;
   justify-content: space-between;
   align-items: center;
+  font-size: small;
+  font-weight: bold;
 }
 
 ul {
